+++ application/features/bootstrap/MfaRateLimitContext.php
@@ -7,7 +7,6 @@ use common\models\EmailLog;
 use common\models\Mfa;
 use common\models\MfaFailedAttempt;
 use common\models\User;
-use Sil\SilIdBroker\Behat\Context\fakes\FakeOfflineLdap;
 use Webmozart\Assert\Assert;
 use Yii;
 use yii\web\TooManyRequestsHttpException;